% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErdosRenyi.R
\name{ErdosRenyi}
\alias{ErdosRenyi}
\title{Generate Erdos-Renyi precision matrix}
\usage{
ErdosRenyi(p, prop, numT)
}
\arguments{
\item{p}{Dimension of network}

\item{prop}{rewiring probability}

\item{numT}{number of timepoints}
}
\value{
A list with components
\item{Precision}{Estimated p by p inverse covariance matrix}
\item{Data}{Data matrix containing the simulated time series for each region (numT by p data matrix)}
}
\description{
Estimate a precision matrix with a Erdos-Renyi network structure and simulate multivariate normal data.
}
\examples{
fit<-ErdosRenyi(40,.15,200)
fit$Precision
fit$Data
}
\references{
Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. \url{http://igraph.org}
}
